code health: centralize utilities, improve backend architecture, and implement toast notifications
- Centralized maskKey, escapeHtml, and escapeRegex in Backend/utils/strings.js
- Extracted OTP, cookie, and AI parsing to dedicated utility files
- Refactored Socket.io into socket.service.js with atomic MongoDB operations
- Modularized CORS and CSRF configurations
- Implemented a centralized logger and replaced all console calls in backend (controllers, services, utils)
- Added JSDoc documentation to all new backend utilities
- Implemented a centralized Toast notification system in the frontend and replaced all alert() calls
- Ensured proper resource cleanup in frontend components
- Verified utility logic and confirmed architectural integrity
