--- Backend/config/constants.js
@@ -1,25 +1,0 @@
-export const PROJECT_CATEGORIES = [
-  'DSA',
-  'Frontend Development',
-  'Backend Development',
-  'Fullstack Development',
-  'Code Review & Optimization',
-  'Testing & QA',
-  'API Development',
-  'Database Engineering',
-  'Software Architecture',
-  'Version Control & Git',
-  'Agile Project Management',
-  'CI/CD Automation',
-  'Debugging & Troubleshooting',
-  'Documentation Generation',
-  'Code Refactoring'
-];
-
-export const ALLOWED_ORIGINS = [
-  'https://chatraj-frontend.vercel.app',
-  'https://chatraj.vercel.app',
-  'http://localhost:5173',
-  'http://localhost:5174',
-  'https://chatraj-fpo1pa3bz-abhiraj-gautams-projects.vercel.app'
-];
