+++ Backend/utils/cookies.js
@@ -1,3 +1,14 @@
+/**
+ * Utilities for handling HTTP cookies in a cross-environment compatible way.
+ */
+
+/**
+ * Extracts an array of individual cookie strings from a 'set-cookie' header.
+ * Handles differences between `node-fetch`, `undici`, and standard Fetch API.
+ *
+ * @param {Response} resp - The HTTP response object.
+ * @returns {string[]} An array of cookie strings.
+ */
 export function extractSetCookieArray(resp) {
   // node-fetch / undici gives headers.raw()
   if (resp.headers && typeof resp.headers.raw === 'function') {
@@ -10,6 +21,13 @@ export function extractSetCookieArray(resp) {
   return String(single).split(/,(?=\s*[^;]+?=)/g);
 }
 
+/**
+ * Builds a 'Cookie' header string from an array of 'set-cookie' strings.
+ * Extracts only the key=value part and joins them with a semicolon.
+ *
+ * @param {string[]} setCookieArray - Array of raw 'set-cookie' strings.
+ * @returns {string} A formatted 'Cookie' header string.
+ */
 export function buildCookieHeader(setCookieArray) {
   if (!Array.isArray(setCookieArray)) return '';
   return setCookieArray.map(c => c.split(';')[0].trim()).filter(Boolean).join('; ');
